renamed directories and files to better represent the system's architecture
