--- server/db/devtest/getResidents.js
@@ -1,6 +1,0 @@
-const data = require('./../model_mysql.js')
-
-data.getResidents('');
-data.getResidents('et');
-data.getResidents('1');
-data.getResidents('a');
