--- server/db/devtest/testdataperm.sql
@@ -1,9 +1,0 @@
-INSERT INTO st_ronans_care.UserPermissions (userID, pmsnID)
-VALUES
-  (9, 1),
-  (9, 2),
-  (9, 3),
-  (9, 4),
-  (8, 2),
-  (8, 3),
-  (8, 4);
