+++ server/persistence/persist_main.js
@@ -1,6 +1,6 @@
 'use strict';
 const mysql = require('mysql2/promise');
-const config = require('./../../config.json');
+const config = require('../../config.json');
 
 let sqlPromise = null;
 async function init() {
